Task_3/requirements.txt
pillow
opencv-python
pytesseract
pandas
python-dateutil
regex
pathway
PyMuPDF
scikit-learn